Refractor code and use mass assignment for consistency

+++ database/migrations/2023_04_18_071042_timelogs.php
@@ -21,7 +21,7 @@ return new class extends Migration
             $table->timestamp('StartBreak')->nullable();
             $table->timestamp('EndBreak')->nullable();
             $table->timestamp('StopWork')->nullable();
-            $table->decimal('BreaksTaken')->default(0);
+            $table->integer('BreaksTaken')->default(0);
             $table->decimal('BreakHours', 5, 2)->default(0, 00);
             $table->decimal('RegularHours', 5, 2)->default(0, 00);
             $table->text('userNote')->nullable();
